eth/tracers: `call_tracer_parity.js` adds compatibility with Parity’s trace module
This tracer, mostly copies `call_tracer.js` with some small additions on `step` and `fault` functions, with actually changing the `result` and `finalize` functions, so as instead of nesting internal transactions under the parent one, it flatters them following the Parity’s formatting.

This way explorers and exchanges using Parity for it’s trace module, can easier migrate to geth and later to its own tracers.

The current version of this file, achieved 100% compatibility with Parity on the ETC’s mordor chain.